% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsectEggs.R
\docType{package}
\name{insect_eggs}
\alias{insect_eggs}
\alias{InsectEggs}
\alias{InsectEggs-package}
\title{InsectEggs}
\format{
A dataframe with 2605 elements
\describe{
\item{ar}{Insect egg aspect ratio (length / width)}
\item{asym}{Insect egg asymmetry}
\item{curv}{Insect egg curvature}
\item{family}{Insect family (taxonomic rank)}
\item{genus}{Insect genus (taxonomic rank)}
\item{name}{Insect name}
\item{order}{Insect order (taxonomic rank)}
\item{species}{Insect species (taxonomic rank)}
\item{vol}{Insect egg volume in mm^3}
\item{year}{Year of publication of a paper from which egg dimensions were extracted}
}
}
\source{
\url{https://www.biorxiv.org/content/10.1101/471953v1}
}
\usage{
insect_eggs
}
\description{
This package contains information about egg
dimensions of insects of various species.

A data set containing information about egg
dimensions of insects of various species.
}
\keyword{datasets}
